+++ lib/pgvector.dart
@@ -1,7 +1,10 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
+import 'halfvec.dart';
 import 'vector.dart';
+
+export 'halfvec.dart' show HalfVector;
 export 'vector.dart' show Vector;
 
 class Pgvector {